requirements.txt
discord.py
mysql-connector-python
requests
python-dotenv
pytest
pytest-asyncio
black